Swap back to normal App component
